requirements.txt
streamlit
pandas
numpy
openai
python-dotenv
chromadb
langchain
langchain-openai
langsmith
matplotlib
seaborn
pyyaml
wordcloud